+++ microservice/utils/src/main/java/org/javaee7/wildfly/samples/everest/utils/exception/InitializationException.java
@@ -14,20 +14,22 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package org.javaee7.wildfly.samples.everest.qualifiers;
-
-import javax.inject.Qualifier;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
+package org.javaee7.wildfly.samples.everest.utils.exception;
 
 /**
  * @author Ryan McGuinness [rmcguinness@walmartlabs.com]
- *         Created: 6/23/15
+ *         Created: 6/19/15
  */
-@Qualifier
-@Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.FIELD, ElementType.PARAMETER, ElementType.METHOD})
-public @interface QServerName {
+public class InitializationException extends RuntimeException {
+    public InitializationException() {
+        super();
+    }
+
+    public InitializationException(String err) {
+        super(err);
+    }
+
+    public InitializationException(Throwable t) {
+        super(t);
+    }
 }
